수정사항 반영

+++ seboard/src/main/java/com/seproject/admin/menu/application/AdminMenuAppService.java
@@ -1,10 +1,16 @@
 package com.seproject.admin.menu.application;
 
+import com.seproject.account.account.domain.Account;
+import com.seproject.account.utils.SecurityUtils;
+import com.seproject.admin.dashboard.domain.DashBoardMenu;
+import com.seproject.admin.dashboard.service.AdminDashBoardServiceImpl;
 import com.seproject.admin.menu.service.AdminMenuService;
 import com.seproject.admin.menu.utils.DelegatingMenuProvider;
 import com.seproject.board.menu.domain.Menu;
 import com.seproject.board.menu.service.MenuService;
 import com.seproject.error.errorCode.ErrorCode;
+import com.seproject.error.exception.CustomAccessDeniedException;
+import com.seproject.error.exception.CustomAuthenticationException;
 import com.seproject.error.exception.CustomIllegalArgumentException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.util.Pair;
@@ -25,8 +31,24 @@ public class AdminMenuAppService {
     private final MenuService menuService;
     private final DelegatingMenuProvider menuProvider;
 
+    private final AdminDashBoardServiceImpl dashBoardService;
+
+    //TODO : 추후 AOP로 변경 필요
+    private void checkAuthorization(){
+        Account account = SecurityUtils.getAccount()
+                .orElseThrow(() -> new CustomAuthenticationException(ErrorCode.NOT_LOGIN, null));
+
+        DashBoardMenu dashboardmenu = dashBoardService.findDashBoardMenuByUrl(DashBoardMenu.MENU_EDIT_URL);
+
+        if(!dashboardmenu.authorize(account.getRoles())){
+            throw new CustomAccessDeniedException(ErrorCode.ACCESS_DENIED, null);
+        }
+    }
+
     @Transactional
     public void createMenu(CreateMenuRequest request, String categoryType) {
+        checkAuthorization();
+
         String name = request.getName();
         if (name == null) {
             throw new CustomIllegalArgumentException(ErrorCode.MENU_NAME_NULL, null);
@@ -36,6 +58,8 @@ public class AdminMenuAppService {
     }
 
     public List<MenuResponse> findAllMenuTree() {
+        checkAuthorization();
+
         //TODO : 성능 개선?
         List<Menu> depth0Menus = menuService.findByDepth(0);
         List<MenuResponse> menuResponses = depth0Menus.stream()
@@ -101,6 +125,8 @@ public class AdminMenuAppService {
     }
 
     public MenuResponse findMenu(Long menuId) {
+        checkAuthorization();
+
         Menu menu = menuService.findById(menuId);
         List<MenuResponse> subMenus = menuService.findSubMenu(menuId)
                 .stream()
@@ -111,17 +137,23 @@ public class AdminMenuAppService {
 
     @Transactional
     public void delete(Long menuId) {
+        checkAuthorization();
+
         adminMenuService.delete(menuId);
     }
 
 
     @Transactional
     public void migrateCategory(Long fromMenuId, Long toMenuId) {
+        checkAuthorization();
+
         adminMenuService.changeSuperCategory(fromMenuId, toMenuId);
     }
 
     @Transactional
     public void update(Long categoryId, UpdateMenuRequest request) {
+        checkAuthorization();
+
         Menu menu = menuService.findById(categoryId);
 
         menuProvider.update(menu, request);
